backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "GraphQL API backend using FastAPI, Strawberry, and SQLAlchemy"
requires-python = ">=3.13,<3.14"
dependencies = [
    "strawberry-graphql[fastapi]>=0.281.0",
    "uvicorn>=0.35.0",
    "sqlalchemy>=2.0.43",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.4",
    "redis>=6.4.0",
    "bcrypt>=4.3.0",
    "authlib==1.6.6",
    "httpx>=0.28.1",
    "structlog>=25.5.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "python-multipart>=0.0.21",
    "miniopy-async>=1.23.4",
    "geoalchemy2>=0.18.1",
    "shapely>=2.1.2",
    "alembic-postgresql-enum>=1.8.0",
    "pycountry>=24.6.1",
    "pillow>=12.1.0",
    "avinit>=1.3.0",
]

[dependency-groups]
dev = [
    "gql>=4.0.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
migration = [
    "alembic>=1.15.1",
    "psycopg2-binary>=2.9.10",
    "alembic-postgresql-enum>=1.8.0",
]
production = [
    "uvloop>=0.21.0",
    "gunicorn>=21.2.0",
]

[project.scripts]
server = "app.cli:server"
migrate = "app.cli:migrate"
clean_db = "app.cli:clean_db"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
